analysis-requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
boto3==1.24.56
botocore==1.27.56
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
discord.py==2.0.0
Django==4.1
django-extensions==3.2.0
django-storages==1.13.1
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
fonttools==4.37.1
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.14.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.1
s3transfer==0.6.0
scipy==1.9.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.4.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
tzdata==2022.2
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
widgetsnbextension==4.0.2
yarl==1.8.1
